Move point in time to server (#70704)
This change moves the implementation of point in time to the server package.
