+++ rest-api-spec/src/main/resources/rest-api-spec/api/close_point_in_time.json
