+++ rest-api-spec/src/main/resources/rest-api-spec/api/open_point_in_time.json
