+++ x-pack/plugin/data-streams/qa/rest/build.gradle
@@ -5,8 +5,8 @@ apply plugin: 'elasticsearch.java-rest-test'
 
 restResources {
   restApi {
-    includeCore 'bulk', 'count', 'search', '_common', 'indices', 'index', 'cluster', 'rank_eval', 'reindex', 'update_by_query', 'delete_by_query'
-    includeXpack 'eql', 'indices', 'data_stream', 'migration', 'async_search', 'searchable_snapshots', 'rollup', 'graph', 'ilm', 'open_point_in_time', 'close_point_in_time'
+    includeCore 'bulk', 'count', 'search', '_common', 'indices', 'index', 'cluster', 'rank_eval', 'reindex', 'update_by_query', 'delete_by_query', 'open_point_in_time', 'close_point_in_time'
+    includeXpack 'eql', 'indices', 'data_stream', 'migration', 'async_search', 'searchable_snapshots', 'rollup', 'graph', 'ilm'
   }
   restTests {
     includeXpack 'data_stream'
